requirements.txt
Flask
Flask-Cors
pandas
sentence-transformers==2.7.0
faiss-cpu
transformers==4.41.2
tokenizers==0.19.1
torch
nltk
accelerate
protobuf==4.25.3
huggingface-hub>=0.23.0
beautifulsoup4
requests
selenium
webdriver-manager